refactor: MainDetailViewController에서 PopupTitleCollectionViewCell 관련 코드 이동
