+++ Popcorn-iOS/Source/Presentation/Main/PopupDetailView/PopupDetailInfoView.swift
@@ -216,39 +216,3 @@ extension PopupDetailInfoView {
         ])
     }
 }
-
-#if canImport(SwiftUI) && DEBUG
-import SwiftUI
-struct UIViewPreview<View: UIView>: UIViewRepresentable {
-    let view: View
-
-    init(_ builder: @escaping () -> View) {
-        view = builder()
-    }
-
-    // MARK: - UIViewRepresentable
-
-    func makeUIView(context: Context) -> UIView {
-        return view
-    }
-
-    func updateUIView(_ view: UIView, context: Context) {
-        view.setContentHuggingPriority(.defaultHigh, for: .horizontal)
-        view.setContentHuggingPriority(.defaultHigh, for: .vertical)
-    }
-}
-
-#endif
-
-#if canImport(SwiftUI) && DEBUG
-import SwiftUI
-
-struct Preview: PreviewProvider {
-    static var previews: some View {
-        UIViewPreview {
-            let myView = PopupDetailInfoView()
-            return myView
-        }.previewLayout(.sizeThatFits)
-    }
-}
-#endif
